requirements.txt

openai
python-telegram-bot==13.15
beautifulsoup4
requests
